+++ internal/product/db/migration/00003_add_idempotency.up.sql
@@ -1,7 +1,7 @@
 CREATE TABLE "idempotency" (
     "id" UUID NOT NULL,
     "product_id" UUID NOT NULL,
-    "quantity" INTEGER NOT NULL CHECK (quantity >= 0),
+    "quantity" BIGINT NOT NULL CHECK (quantity >= 0),
     "rollbacked" BOOLEAN NOT NULL DEFAULT FALSE,
     "created_at" TIMESTAMPTZ NOT NULL DEFAULT (now()),
     PRIMARY KEY ("id" , "product_id")
